refactor+fix:(Order): Change order quantity from double to Integer and fix total amount rounding
- Updated quantity type from double to Integer in OrderLine model, DTOs (request/response)
- Changed database schema: order_line.quantity column from DOUBLE to INT
- Added BigDecimal scale and rounding (2 decimals, HALF_UP) in OrderTotalCalculator
- Fixed product-service URL in application.yml
- Added order_status column with NOT NULL constraint in orders table migration

+++ services/order-service/src/main/java/com/yamar/orderservice/dto/OrderLineResponse.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 public record OrderLineResponse(
         Long id,
-        double quantity,
+        Integer quantity,
         String productId,
         BigDecimal pricePerUnit,
         BigDecimal subTotal
